QCalculator to version 2
Changes in this version:

Program's outlook is redesigned.

Full Keypad/Keypress support is added. 

Notice: Back key is Space. Return key is equals.

If your keys do not see to work, click the display part of the calculator to remedy the issue.
Also make sure that you have numlock enabled if you are using keypad.